Add API implementation for auth service (#3178)

+++ pkg/auth/authenticator.go
@@ -22,7 +22,7 @@ import (
 type Authenticator interface {
 	// AuthenticateUser authenticates a user matching username and
 	// password and returns their ID.
-	AuthenticateUser(ctx context.Context, username, password string) (int, error)
+	AuthenticateUser(ctx context.Context, username, password string) (int64, error)
 }
 
 // Credentialler fetches S3-style credentials for access keys.
@@ -40,7 +40,7 @@ func NewChainAuthenticator(auth ...Authenticator) Authenticator {
 // order, returning the last error in case all fail.
 type ChainAuthenticator []Authenticator
 
-func (ca ChainAuthenticator) AuthenticateUser(ctx context.Context, username, password string) (int, error) {
+func (ca ChainAuthenticator) AuthenticateUser(ctx context.Context, username, password string) (int64, error) {
 	var merr *multierror.Error
 	logger := logging.FromContext(ctx).WithField("username", username)
 	for _, a := range ca {
@@ -63,7 +63,7 @@ func NewEmailAuthenticator(service Service) *EmailAuthenticator {
 	return &EmailAuthenticator{AuthService: service}
 }
 
-func (e EmailAuthenticator) AuthenticateUser(ctx context.Context, username, password string) (int, error) {
+func (e EmailAuthenticator) AuthenticateUser(ctx context.Context, username, password string) (int64, error) {
 	user, err := e.AuthService.GetUserByEmail(ctx, username)
 	if err != nil {
 		return InvalidUserID, err
@@ -90,7 +90,7 @@ func NewBuiltinAuthenticator(service Service) *BuiltinAuthenticator {
 	return &BuiltinAuthenticator{creds: service}
 }
 
-func (ba *BuiltinAuthenticator) AuthenticateUser(ctx context.Context, username, password string) (int, error) {
+func (ba *BuiltinAuthenticator) AuthenticateUser(ctx context.Context, username, password string) (int64, error) {
 	// Look user up in DB.  username is really the access key ID.
 	cred, err := ba.creds.GetCredentials(ctx, username)
 	if err != nil {
@@ -162,7 +162,7 @@ func inBrackets(filter string) string {
 	return fmt.Sprintf("(%s)", filter)
 }
 
-func (la *LDAPAuthenticator) AuthenticateUser(ctx context.Context, username, password string) (int, error) {
+func (la *LDAPAuthenticator) AuthenticateUser(ctx context.Context, username, password string) (int64, error) {
 	// There may be multiple authenticators.  Log everything to allow debugging.
 	logger := logging.FromContext(ctx).WithField("username", username)
 	controlConn, err := la.getControlConnection(ctx)
